#910 Change: extract billing and pricing feature
