+++ src/features/billing/components/PricingSection.tsx
@@ -8,11 +8,11 @@ import {
     PRICING_TIERS,
     type BillingInterval,
     type PricingTier,
-} from "@/components/landing-page/pricing-section/PricingSection.data.ts";
+} from "@/features/billing/components/PricingSection.data.ts";
 import { useUserPreferences } from "@/features/preferences/hooks/useUserPreferences.tsx";
 import { Link } from "@tanstack/react-router";
 import { useState } from "react";
-import { useStripeBilling } from "@/hooks/billing/useStripeBilling.ts";
+import { useStripeBilling } from "@/features/billing/hooks/useStripeBilling.ts";
 import type { BillingCycle } from "@/data/internal/billing/BillingCycle.ts";
 import type { BillingPlan } from "@/data/internal/billing/BillingPlan.ts";
 import { useResolvedAuth } from "@/features/authentication/hooks/useResolvedAuth.ts";
