+++ src/routes/$lng.index.tsx
@@ -4,7 +4,7 @@ import FAQSection from "@/components/landing-page/faq-section/FAQSection.tsx";
 import HeroSection from "@/components/landing-page/hero-section/HeroSection.tsx";
 import FeaturesSection from "@/components/landing-page/features-section/FeaturesSection.tsx";
 import HowItWorksSection from "@/components/landing-page/how-it-works-section/HowItWorksSection.tsx";
-import PricingSection from "@/components/landing-page/pricing-section/PricingSection.tsx";
+import PricingSection from "@/features/billing/components/PricingSection.tsx";
 import NewsletterSection from "@/components/landing-page/newsletter-section/NewsletterSection.tsx";
 import { createFileRoute } from "@tanstack/react-router";
 import { generatePageHeadMeta } from "@/lib/seo/pageHeadMeta.ts";
